+++ src/main/java/com/wormhole_xtreme/wormhole/model/StargateManager.java
@@ -1192,7 +1192,7 @@ public class StargateManager
      */
     public static boolean isPortalBlock(final Block b)
     {
-        if (b == null || b.getWorld() == null)
+        if (b == null)
         {
             return false;
         }